# apps/fx-node/openapi.yaml
openapi: 3.1.0
info:
  title: Blueballs FX Node
  version: 0.3.0-reference
  description: |
    Canonical REST contract for the self-hostable Blueballs FX reference runtime.

    The `reference-sandbox` composes institution policy, signed private liquidity,
    issuer and institutional inventory, reference pricing, principal risk,
    multi-source reservation, fiat settlement state and the public BRL → EUR
    reference trade and a sandbox Monetary Engine for reserve-backed issuance,
    purpose-bound settlement receipts and transparent remittance pricing.
    Preview endpoints do not reserve. A firm sandbox quote is returned only
    after every selected source reserves capacity. Monetary instruments are
    simulations, not production assets or claims on Blueballs.
  license:
    name: MIT
    identifier: MIT
servers:
  - url: http://localhost:8788
    description: Local reference sandbox
security:
  - bearerAuth: []
...
paths:
...
  /v2/fx/reference/status:
    get:
      tags: [Runtime]
      security: []
      summary: Inspect the canonical reference composition
      responses:
        '200':
          description: Runtime capabilities and evidence boundary
          content:
            application/json:
              schema: { type: object, additionalProperties: true }
        default: { $ref: '#/components/responses/Error' }
  /v2/fx/reference/policy:
    get:
      tags: [Runtime]
      security: []
      summary: Inspect active institution policy
      responses:
        '200':
          description: Policy summary
          content:
            application/json:
              schema: { type: object, additionalProperties: true }
        default: { $ref: '#/components/responses/Error' }
  /v2/fx/reference/market:
    get:
      tags: [Runtime]
      security: []
      summary: Inspect the current public-reference market state
      responses:
        '200':
          description: Current scenario, eligible sources and reference-price state
          content:
            application/json:
              schema: { $ref: '#/components/schemas/ReferenceMarket' }
        default: { $ref: '#/components/responses/Error' }
  /v2/fx/reference/scenario:
    get:
      tags: [Runtime]
      security: []
      summary: List scenarios and inspect the active one
      responses:
        '200':
          description: Scenario catalogue and current market state
          content:
            application/json:
              schema: { type: object, additionalProperties: true }
        default: { $ref: '#/components/responses/Error' }
    post:
      tags: [Runtime]
      summary: Apply a deterministic backend market scenario
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [id]
              properties:
                id:
                  type: string
                  enum: [balanced, lp_offline, issuer_policy_blocked, treasury_near_limit, principal_limit, reference_outage]
      responses:
        '200':
          description: Updated market state
          content:
            application/json:
              schema: { $ref: '#/components/schemas/ReferenceMarket' }
        default: { $ref: '#/components/responses/Error' }
  /v2/fx/reference/liquidity:
    get:
      tags: [Liquidity]
      security: []
      summary: Inspect currently policy-eligible source slices without reserving them
      parameters:
        - $ref: '#/components/parameters/InputAsset'
        - $ref: '#/components/parameters/OutputAsset'
        - name: exactOutput
          in: query
          schema: { $ref: '#/components/schemas/AtomicAmount' }
      responses:
        '200':
          description: Eligible source slices
          content:
            application/json:
              schema:
                type: object
                required: [object, data]
                properties:
                  object: { type: string, const: list }
                  data:
                    type: array
                    items: { $ref: '#/components/schemas/LiquiditySlice' }
        default: { $ref: '#/components/responses/Error' }
  /v2/fx/reference/settlement-route:
    get:
      tags: [Runtime]
      security: []
      summary: Inspect the seeded BRL → internal deposit claim → EURC → EUR route
      responses:
        '200':
          description: Mixed-finality settlement graph analysis
          content:
            application/json:
              schema: { $ref: '#/components/schemas/SettlementRoute' }
        default: { $ref: '#/components/responses/Error' }
...
  /v2/fx/reference/monetary/instruments:
    get:
      tags: [Monetary Engine]
      security: []
      summary: List configured reference instruments
      responses:
        '200':
          description: Stablecoin and tokenized-deposit reference instruments
          content:
            application/json:
              schema:
                type: object
                required: [object, data]
                properties:
                  object: { type: string, const: list }
                  data:
                    type: array
                    items: { $ref: '#/components/schemas/MonetaryInstrument' }
        default: { $ref: '#/components/responses/Error' }
  /v2/fx/reference/monetary/remittance/preview:
    post:
      tags: [Monetary Engine]
      security: []
      summary: Preview the canonical BRL to EUR monetary corridor
      description: Compatibility endpoint that delegates to the canonical policy-approved liquidity pipeline. Reference observations are deterministic valuation fixtures, not external or executable oracle prices. It never reserves or moves value.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [inputAmount]
              properties:
                inputAmount: { $ref: '#/components/schemas/AtomicAmount' }
                inputCurrency: { type: string, const: BRL, default: BRL }
                outputCurrency: { type: string, const: EUR, default: EUR }
      responses:
        '200':
          description: Indicative unreserved BRL to EUR preview
          content:
            application/json:
              schema: { $ref: '#/components/schemas/PublicTrade' }
        '409': { $ref: '#/components/responses/Error' }
        default: { $ref: '#/components/responses/Error' }
  /v2/fx/reference/monetary/reserves:
    post:
      tags: [Monetary Engine]
      summary: Record an unsettled external reserve deposit
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [reserveCurrency, amount, providerRef]
              properties:
                reserveCurrency: { type: string }
                amount: { $ref: '#/components/schemas/AtomicAmount' }
                providerRef: { type: string, description: Idempotency reference from the reserve provider. }
      responses:
        '201':
          description: Pending reserve deposit; it cannot support issuance yet
          content:
            application/json:
              schema: { $ref: '#/components/schemas/ReserveDeposit' }
        '409': { $ref: '#/components/responses/Error' }
        default: { $ref: '#/components/responses/Error' }
  /v2/fx/reference/monetary/reserves/{depositId}:
    get:
      tags: [Monetary Engine]
      summary: Retrieve one reserve deposit
      parameters:
        - $ref: '#/components/parameters/DepositId'
      responses:
        '200':
          description: Reserve deposit
          content:
            application/json:
              schema: { $ref: '#/components/schemas/ReserveDeposit' }
        '404': { $ref: '#/components/responses/Error' }
  /v2/fx/reference/monetary/reserves/{depositId}/settle:
    post:
      tags: [Monetary Engine]
      summary: Mark external reserve funds as irrevocably settled
      parameters:
        - $ref: '#/components/parameters/DepositId'
      responses:
        '200':
          description: Settled reserve deposit now eligible to support issuance or a receipt
          content:
            application/json:
              schema: { $ref: '#/components/schemas/ReserveDeposit' }
        '409': { $ref: '#/components/responses/Error' }
        default: { $ref: '#/components/responses/Error' }
  /v2/fx/reference/monetary/risk-capital:
    post:
      tags: [Monetary Engine]
      summary: Set separately funded FX risk capital
      description: Risk capital is reported but never counted as issuance reserve.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [reserveCurrency, amount]
              properties:
                reserveCurrency: { type: string }
                amount: { type: string, pattern: '^\\d+$' }
      responses:
        '200': { description: Updated isolated risk-capital position }
        default: { $ref: '#/components/responses/Error' }
  /v2/fx/reference/monetary/instruments/{instrumentCode}/mint:
    post:
      tags: [Monetary Engine]
      summary: Mint a sandbox instrument against settled unallocated reserve
      parameters:
        - $ref: '#/components/parameters/InstrumentCode'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [amount, beneficiaryRef]
              properties:
                amount: { $ref: '#/components/schemas/AtomicAmount' }
                beneficiaryRef: { type: string }
      responses:
        '201': { description: Sandbox issuance event }
        '409': { $ref: '#/components/responses/Error' }
        default: { $ref: '#/components/responses/Error' }
  /v2/fx/reference/monetary/instruments/{instrumentCode}/redeem:
    post:
      tags: [Monetary Engine]
      summary: Burn sandbox supply and pay from settled reserve
      parameters:
        - $ref: '#/components/parameters/InstrumentCode'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [amount, holderRef]
              properties:
                amount: { $ref: '#/components/schemas/AtomicAmount' }
                holderRef: { type: string }
      responses:
        '200': { description: Sandbox redemption event }
        '409': { $ref: '#/components/responses/Error' }
        default: { $ref: '#/components/responses/Error' }
  /v2/fx/reference/monetary/receipts:
    post:
      tags: [Monetary Engine]
      summary: Create an expiring, purpose-bound settlement receipt
      description: Receipts are non-transferable escrow claims, not money or stablecoins.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [reserveCurrency, amount, beneficiaryRef, purpose, expiresAt]
              properties:
                reserveCurrency: { type: string }
                amount: { $ref: '#/components/schemas/AtomicAmount' }
                beneficiaryRef: { type: string }
                purpose: { type: string }
                expiresAt: { type: integer }
      responses:
        '201':
          description: Active receipt with reserve locked until consume or expiry
          content:
            application/json:
              schema: { $ref: '#/components/schemas/SettlementReceipt' }
        '409': { $ref: '#/components/responses/Error' }
        default: { $ref: '#/components/responses/Error' }
  /v2/fx/reference/monetary/receipts/{receiptId}:
    get:
      tags: [Monetary Engine]
      summary: Retrieve one settlement receipt
      parameters:
        - $ref: '#/components/parameters/ReceiptId'
      responses:
        '200':
          description: Settlement receipt
          content:
            application/json:
              schema: { $ref: '#/components/schemas/SettlementReceipt' }
        '404': { $ref: '#/components/responses/Error' }
...
  /v2/fx/reference/monetary/events:
    get:
      tags: [Monetary Engine]
      summary: List monetary state-transition evidence
      parameters:
        - name: limit
          in: query
          schema: { type: integer, minimum: 1, maximum: 200, default: 50 }
      responses:
        '200': { description: Monetary event list }
        default: { $ref: '#/components/responses/Error' }
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
  parameters:
    InputAsset:
      name: inputAsset
      in: query
      required: true
      schema: { type: string }
    OutputAsset:
      name: outputAsset
      in: query
      required: true
      schema: { type: string }
...
  schemas:
...
    ReferenceMarket:
      type: object
      required: [id, sources, reference]
      properties:
        id: { type: string }
        sources:
          type: array
          items: { $ref: '#/components/schemas/SourceStatus' }
        reference: { type: object, additionalProperties: true }
...
  responses:
    Error:
      description: Machine-readable FX node error
      content:
        application/json:
          schema: { $ref: '#/components/schemas/Error' }